+++ android/src/main/kotlin/be/tramckrijte/workmanager/WorkmanagerPlugin.kt
@@ -1,23 +1,38 @@
 package be.tramckrijte.workmanager
 
+import android.content.Context
+import io.flutter.embedding.android.FlutterEngineConfigurator
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry
-import io.flutter.plugin.common.PluginRegistry.Registrar
 
-class WorkmanagerPlugin(private val workmanagerCallHandler: WorkmanagerCallHandler) : MethodCallHandler {
+class WorkmanagerPlugin : MethodCallHandler, FlutterPlugin {
+
+    private lateinit var workmanagerCallHandler: WorkmanagerCallHandler
+
+    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) =
+            registerWorkManager(binding.binaryMessenger, binding.applicationContext)
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {}
 
     companion object {
-        lateinit var pluginRegistryCallback: PluginRegistry.PluginRegistrantCallback
+        var pluginRegistryCallback: PluginRegistry.PluginRegistrantCallback? = null
 
         @JvmStatic
-        fun registerWith(registrar: Registrar) {
-            val channel = MethodChannel(registrar.messenger(), "be.tramckrijte.workmanager/foreground_channel_work_manager")
-            channel.setMethodCallHandler(WorkmanagerPlugin(WorkmanagerCallHandler(registrar.activeContext())))
+        private fun registerWorkManager(messenger: BinaryMessenger, ctx: Context) {
+            val channel = MethodChannel(messenger, "be.tramckrijte.workmanager/foreground_channel_work_manager")
+            channel.setMethodCallHandler(WorkmanagerPlugin().apply { workmanagerCallHandler = WorkmanagerCallHandler(ctx) })
         }
 
+        @JvmStatic
+        fun registerWith(registrar: PluginRegistry.Registrar) =
+                registerWorkManager(registrar.messenger(), registrar.activeContext())
+
+        @Deprecated(message = "Use the Android v2 embedding method.")
         @JvmStatic
         fun setPluginRegistrantCallback(pluginRegistryCallback: PluginRegistry.PluginRegistrantCallback) {
             WorkmanagerPlugin.pluginRegistryCallback = pluginRegistryCallback