+++ example/android/app/src/main/kotlin/be/tramckrijte/workmanager_example/EmbeddingV1Application.kt
@@ -5,13 +5,14 @@ import io.flutter.app.FlutterApplication
 import io.flutter.plugin.common.PluginRegistry
 import io.flutter.plugins.GeneratedPluginRegistrant
 
-class App : FlutterApplication(), PluginRegistry.PluginRegistrantCallback {
+@Deprecated(message = "Not used, but here to show you how you can use the plugin using the old v1 embedding method.")
+class EmbeddingV1Application : FlutterApplication(), PluginRegistry.PluginRegistrantCallback {
+    override fun registerWith(registry: PluginRegistry) {
+        GeneratedPluginRegistrant.registerWith(registry)
+    }
+
     override fun onCreate() {
         super.onCreate()
         WorkmanagerPlugin.setPluginRegistrantCallback(this)
-    }
-
-    override fun registerWith(reg: PluginRegistry?) {
-        GeneratedPluginRegistrant.registerWith(reg)
     }
 }