+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+* Android:
+    * Adding support for the `Android v2 embedding`.  
+      See more details in the Android setup guide.
+    
 # 0.1.5
  * Android:
     * The result was not correctly mapped to a RETRY when returning false from a background task